BinarySearch/BinarySearch_test.go
/**
 * @Author: zhangxuesong(joseph2018@aliyun.com)
 * @Date: 2022/6/21
 * @Description:
 */
package BinarySearch

import "testing"

func TestSearch(t *testing.T) {
	type args struct {
		nums   []int
		target int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{name: "1", args: args{nums: []int{-1, 0, 3, 5, 9, 12}, target: 9}, want: 4},
		{name: "2", args: args{nums: []int{-1, 0, 3, 5, 9, 12}, target: 2}, want: -1},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := Search(tt.args.nums, tt.args.target); got != tt.want {
				t.Errorf("Search() = %v, want %v", got, tt.want)
			}
		})
	}
}
